+++ src/main/java/com/docket/cartorio/repositories/CartorioRepository.java
@@ -3,5 +3,5 @@ package com.docket.cartorio.repositories;
 import com.docket.cartorio.entities.Cartorio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartorioRepository extends JpaRepository<Cartorio, Integer> {
+public interface CartorioRepository extends JpaRepository<Cartorio, Long> {
 }
